fix: move TokenActionTypes out to util
